\name{pheno}
\alias{pheno}
\title{A longitudinal data with diagnostic results for pre-determined time}
\description{A longitudinal data (629 x 3) with diagnostic results for 4 pre-determined times.
}
\usage{data(pheno)}
\format{A data frame with 629 observations on 3 variables.
  \itemize{
  \item ID: subject IDs
  \item time: Pre-determined diagnostic times. (1-4)
  \item result: diagnostic results (0=No, 1=Yes)
}}
\keyword{datasets}
